re-org with tests.
