+++ scripts/testng-junit/setup.cfg
@@ -1,0 +1,5 @@
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts = --verbose
