+++ scripts/testng-junit/src/tests/setup.py
@@ -1,0 +1,12 @@
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import testng2junit
+
+
+###
+# Strip content for equality comparison.
+###
+def assert_equal_content(content, expected):
+    assert ''.join(content.split()) == ''.join(expected.split())
